requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
ollama==0.5.1
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-slugify==8.0.4
rdflib==7.1.4
requests==2.32.4
sniffio==1.3.1
sparqlwrapper==2.0.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
